Added HTTP Get to agent management (Backend Integration)
