+++ src/app/agent-management/agent-management.component.html
@@ -4,8 +4,8 @@
       <button class="btn btn-primary w-100" id="button-add-agent" type="button"><i class="bi bi-plus"></i>Add Agent</button>
       <table class="table">
         <tbody>
-          <tr>
-            <td class="text-start pt-3"> Michael Jordan</td>
+          <tr *ngFor="let agent of this.agents">
+            <td class="text-start pt-3"> {{agent.name}}</td>
             <td><button type="button" class="btn btn-secondary float-end btn-light">View</button></td>
           </tr>
         </tbody>
