--- src/app/agent-management/agent-management/agent-management.component.ts
@@ -1,12 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  selector: 'app-agent-management',
-  standalone: true,
-  imports: [],
-  templateUrl: './agent-management.component.html',
-  styleUrl: './agent-management.component.css'
-})
-export class AgentManagementComponent {
-
-}
