+++ src/app/app.routes.ts
@@ -3,7 +3,7 @@ import { HomeComponent } from './home/home.component';
 import { RealEstateManagementComponent } from './real-estate-management/real-estate-management.component';
 import { ReferenceDataComponent } from './reference-data/reference-data.component';
 import { CustomerManagementComponent } from './customer-management/customer-management.component';
-import { AgentManagementComponent } from './agent-management/agent-management/agent-management.component';
+import { AgentManagementComponent } from './agent-management/agent-management.component';
 
 const routerConfig: Routes = [
     {
